init/requirements.txt
avro-python3==1.10.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
flask-sock==0.6.0
grpcio==1.51.1
grpcio-tools==1.51.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==4.21.12
requests==2.28.2
simple-websocket==0.9.0
urllib3==1.26.14
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0